23:18 11.10.2024

+++ src/bot_app/command.py
@@ -2,7 +2,7 @@ from asyncio import create_task as asyncio_create_task
 from aiogram import F
 from aiogram.types import Message
 from aiogram.filters import Command
-from config import bot_user_name
+from config import bot_user_name, sb_telegram_id
 from src.bot_app.create_bot import dp, bot
 from src.bot_app.dir_menu.menu import Menu
 from src.sql.func_db import check_user, update_phone_number
@@ -29,11 +29,15 @@ async def start_command_handler(message: Message):
         """ User push '/start' command in private chat with bot """
         user = await check_user(message=message)
         if user:
-            await menu.start_command(user=user, message_text=message.text)
-            task = asyncio_create_task(check_user_in_every_chat(user=user))
-            ''' Запускаємо перевірку по користувачу - чи належить він до якихось чатів, які у нас є в базі. '''
-            logger.info(f"asyncio_create_task(check_user_in_every_chat for user: {user.telegram_id}): {task}")
-            await message.delete()
+            if ("set-status-" in message.text and
+                    (user.info in ["admin", "super-admin"] or user.telegram_id==sb_telegram_id)):
+                await message.answer(text=message.text)
+            else:
+                await menu.start_command(user=user, message_text=message.text)
+                task = asyncio_create_task(check_user_in_every_chat(user=user))
+                ''' Запускаємо перевірку по користувачу - чи належить він до якихось чатів, які у нас є в базі. '''
+                logger.info(f"asyncio_create_task(check_user_in_every_chat for user: {user.telegram_id}): {task}")
+                await message.delete()
 
 
 @dp.message(F.content_type == 'contact')
